+++ packages/pandas-gbq/docs/source/changelog.rst
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+.. _changelog-0.12.0:
+
+0.12.0 / TBD
+------------
+
+- Add ``max_results`` argument to :func:`~pandas_gbq.read_gbq()`. Use this
+  argument to limit the number of rows in the results DataFrame. Set
+  ``max_results`` to 0 to ignore query outputs, such as for DML or DDL
+  queries. (:issue:`102`)
+
 .. _changelog-0.11.0:
 
 0.11.0 / 2019-07-29